# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Al-Tayebat - Healthy Food & Keto Products API for Jordan
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders (guest or user)
      parameters:
        - name: sessionId
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Place a new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Created order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found
...
components:
  schemas:
...
    Order:
      type: object
      required:
        [
          id,
          status,
          total,
          deliveryFee,
          subtotal,
          createdAt,
          items,
          deliveryAddress,
        ]
      properties:
        id:
          type: integer
        sessionId:
          type: ["string", "null"]
        status:
          type: string
          enum:
            [
              pending,
              confirmed,
              preparing,
              ready,
              out_for_delivery,
              on_the_way,
              delivered,
              cancelled,
            ]
        fulfillmentType:
          type: string
          enum: [delivery, pickup]
        vendorId:
          type: ["integer", "null"]
        subtotal:
          type: number
        deliveryFee:
          type: number
        total:
          type: number
        deliveryAddress:
          type: string
        customerName:
          type: ["string", "null"]
        customerPhone:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        estimatedDelivery:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
    OrderItem:
      type: object
      required:
        [
          id,
          productId,
          productNameAr,
          productName,
          quantity,
          unitPrice,
          totalPrice,
        ]
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        productNameAr:
          type: string
        productImageUrl:
          type: ["string", "null"]
        quantity:
          type: integer
        unitPrice:
          type: number
        totalPrice:
          type: number
...
    OrderInput:
      type: object
      required: [deliveryAddress, customerName, customerPhone]
      properties:
        sessionId:
          type: string
        deliveryAddress:
          type: string
        customerName:
          type: string
        customerPhone:
          type: string
        notes:
          type: string
        fulfillmentType:
          type: string
          enum: [delivery, pickup]
        paymentMethod:
          type: string
        userId:
          type: integer
...